# improve `update` performance