--- a/codebase2/codebase-sqlite/sql/018-add-derived-dependents-by-dependency-index.sql
+++ b/codebase2/codebase-sqlite/sql/018-add-derived-dependents-by-dependency-index.sql
@@ -1,0 +1,2 @@
+CREATE INDEX derived_dependents_by_dependency
+  ON dependents_index (dependency_object_id, dependency_component_index);