--- a/parser-typechecker/src/Unison/Codebase/SqliteCodebase/Migrations.hs
+++ b/parser-typechecker/src/Unison/Codebase/SqliteCodebase/Migrations.hs
@@ -87,7 +87,8 @@ migrations regionVar getDeclType termBuffer declBuffer rootCodebasePath =
       (17 {- This migration takes a raw sqlite connection -}, \conn -> migrateSchema16To17 conn),
       sqlMigration 18 Q.addProjectBranchLastAccessedColumn,
       sqlMigration 19 Q.addMergeBranchTables,
-      sqlMigration 20 Q.addUpdateBranchTable
+      sqlMigration 20 Q.addUpdateBranchTable,
+      sqlMigration 21 Q.addDerivedDependentsByDependencyIndex
     ]
   where
     runT :: Sqlite.Transaction () -> Sqlite.Connection -> IO ()