--- a/unison-cli/src/Unison/Cli/UpdateUtils.hs
+++ b/unison-cli/src/Unison/Cli/UpdateUtils.hs
@@ -20,7 +20,7 @@ module Unison.Cli.UpdateUtils
 where
 
 import Control.Monad.Reader (ask)
-import Data.Bifoldable (bifold, bifoldMap)
+import Data.Bifoldable (bifold)
 import Data.Bitraversable (bitraverse)
 import Data.Foldable qualified as Foldable
 import Data.List qualified as List
@@ -157,7 +157,7 @@ getNamespaceDependentsOf2 ::
 getNamespaceDependentsOf2 defns dependencies = do
   let toTermScope = Set.mapMaybe Referent.toReferenceId . BiMultimap.dom
   let toTypeScope = Set.mapMaybe Reference.toId . BiMultimap.dom
-  let scope = bifoldMap toTermScope toTypeScope defns
+  let scope = bimap toTermScope toTypeScope defns
   Operations.transitiveDependentsWithinScope scope dependencies
     <&> bimap (Set.foldl' addTerms Map.empty) (Set.foldl' addTypes Map.empty)
   where
@@ -180,7 +180,7 @@ getNamespaceDependentsOf3 ::
 getNamespaceDependentsOf3 defns dependencies = do
   let toTermScope = Set.mapMaybe Referent.toReferenceId . BiMultimap.dom
   let toTypeScope = Set.mapMaybe Reference.toId . BiMultimap.dom
-  let scope = bifoldMap toTermScope toTypeScope defns
+  let scope = bimap toTermScope toTypeScope defns
   Operations.transitiveDependentsWithinScope scope (bifold dependencies)
 
 ------------------------------------------------------------------------------------------------------------------------