## Overview

This PR improves `update` performance in two ways:

- For `add`-only updates, we skip identifying transitive dependents of updates. (Previously we'd do a big query that'd always return nothing).
- I optimized the query to identify transitive dependents within a scope. SQLite apparently doesn't really know anything about temporary table statistics and was picking a bad join order between `dependents_index` (3.5M rows on my machine) and the temporary scope table (~7k in base). I tried a few tricks but the only one that worked is simulating an inner join with a row-valued `IN` clause. I also added a useful index for the search.

Couple perf numbers:
 
- For an update with 900 transitive dependents, this PR dropped time to identify them from 3870 ms → 169 ms
- For an add-only update, this PR dropped time to identity the (zero) transitive dependents from 673 ms → 0 ms (we skip that)

## Loose ends

`update` can be made faster with more effort. Parsing, rendering, typechecking, and even simple things like saving updated things to the codebase and updating the root branch are all part of the update process and all take a noticeable amount of time.