@aryairani 2025-08-28
> Some tests with 'continue-on-error: true' have failed:
> 
> * Cabal / smoke-test
> 
> _Created by [continue-on-error-comment](https://github.com/mainmatter/continue-on-error-comment/)_

Hey it worked! Or at least did something. cc @sellout