@sellout 2025-08-28
> > Some tests with 'continue-on-error: true' have failed:
>
> It looks like this will probably trigger now whenever there's a build error (i.e. neither `stack` or `cabal` can build). cc @sellout

Hrmm, yeah, maybe we can run that workflow only if the `build-ucm` job passes.